fix: directory and path
