--- experiment/pir/pps/simple/BUILD.bazel
@@ -1,60 +1,0 @@
-cc_library(
-    name = "generate_rand",
-    srcs = ["include/generate_rand.cc"],
-    hdrs = ["include/generate_rand.h"],
-    includes = ["include"],
-    visibility = ["//visibility:public"],
-)
-
-cc_library(
-    name = "data_transmit",
-    srcs = ["include/data_transmit.cc"],
-    hdrs = ["include/data_transmit.h"],
-    includes = ["include"],
-    visibility = ["//visibility:public"],
-)
-
-cc_library(
-    name = "inner_product",
-    srcs = ["include/inner_product.cc"],
-    hdrs = ["include/inner_product.h"],
-    includes = ["include"],
-    visibility = ["//visibility:public"],
-)
-
-cc_library(
-    name = "pir_client",
-    srcs = ["include/pir_client.cc"],
-    hdrs = ["include/pir_client.h"],
-    deps = [
-        ":generate_rand",
-        ":data_transmit",
-        ":inner_product",
-    ],
-    includes = ["include"],
-    visibility = ["//visibility:public"],
-)
-
-cc_library(
-    name = "pir_server",
-    srcs = ["include/pir_server.cc"],
-    hdrs = ["include/pir_server.h"],
-    deps = [
-        ":generate_rand",
-        ":data_transmit",
-        ":inner_product",
-    ],
-    includes = ["include"],
-    visibility = ["//visibility:public"],
-)
-
-cc_binary(
-    name = "pir_test",
-    srcs = ["pir_test.cc"],
-    deps = [
-        ":pir_client",
-        ":pir_server",
-    ],
-    linkopts = ["-pthread"],
-    copts = ["-g"],
-)
