+++ experiment/pir/pps/simplepir/BUILD.bazel
@@ -1,0 +1,69 @@
+# Copyright 2025 The secretflow authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+cc_library(
+    name = "generate_rand",
+    srcs = ["generate_rand.cc"],
+    hdrs = ["generate_rand.h"],
+    visibility = ["//visibility:public"],
+)
+
+cc_library(
+    name = "data_transmit",
+    srcs = ["data_transmit.cc"],
+    hdrs = ["data_transmit.h"],
+    visibility = ["//visibility:public"],
+)
+
+cc_library(
+    name = "inner_product",
+    srcs = ["inner_product.cc"],
+    hdrs = ["inner_product.h"],
+    visibility = ["//visibility:public"],
+)
+
+cc_library(
+    name = "pir_client",
+    srcs = ["pir_client.cc"],
+    hdrs = ["pir_client.h"],
+    deps = [
+        ":generate_rand",
+        ":data_transmit",
+        ":inner_product",
+    ],
+    visibility = ["//visibility:public"],
+)
+
+cc_library(
+    name = "pir_server",
+    srcs = ["pir_server.cc"],
+    hdrs = ["pir_server.h"],
+    deps = [
+        ":generate_rand",
+        ":data_transmit",
+        ":inner_product",
+    ],
+    visibility = ["//visibility:public"],
+)
+
+cc_binary(
+    name = "pir_test",
+    srcs = ["pir_test.cc"],
+    deps = [
+        ":pir_client",
+        ":pir_server",
+    ],
+    linkopts = ["-pthread"],
+    copts = ["-g"],
+)
