+++ experiment/pir/pps/simplepir/pir_client.cc
@@ -22,7 +22,10 @@ namespace pir::simple {
 PIRClient::PIRClient(size_t n, size_t q, size_t N, size_t p, int radius,
                      double sigma, std::string ip, int port)
     : n_(n), q_(q), N_(N), p_(p), ip_(ip), port_(port) {
-  delta_ = q / p;
+  delta_ = static_cast<size_t>(
+      floor(static_cast<double>(q) / static_cast<double>(p)));
+  std::cout << delta_ << std::endl;
+  std::cout << q / p << std::endl;
   precompute_discrete_gaussian(radius, sigma);
 }
 
@@ -95,8 +98,8 @@ void PIRClient::client_answer() {
 void PIRClient::client_recover() {
   __uint128_t d_ =
       ans_[idx_row_] - fast_inner_product_modq(hint_[idx_row_], s_, q_);
-  __uint128_t d =
-      ((d_ % delta_) > (delta_ / 2)) ? (d_ / delta_ + 1) : (d_ / delta_);
+  __uint128_t d = static_cast<__uint128_t>(
+      (d_ % delta_ >= delta_ / 2) ? (d_ / delta_) - 1 : (d_ / delta_));
   std::cout << "Recovered value: " << d % p_ << std::endl;
 }
 }  // namespace pir::simple
